requirements.txt
autopep8==1.6.0
mss==6.1.0
numpy==1.21.4
opencv-python==4.5.4.60
Pillow==8.4.0
PyYAML==6.0
requests==2.26.0
pyclick
telegram
python_telegram_bot==13.8.1
